requirements.txt
anyio>=4.2.0
asgiref>=3.7.2
async-property>=0.2.2
attrs>=23.1.0
autobahn>=23.6.2
Automat>=22.10.0
beautifulsoup4>=4.12.2
certifi>=2023.11.17
cffi>=1.16.0
charset-normalizer>=3.3.2
click>=8.1.7
constantly>=23.10.4
cryptography>=41.0.7
daphne>=4.0.0
Django~=5.1.4
django-cors-headers>=4.3.1
django-debug-toolbar>=4.2.0
django-environ>=0.11.2
django-redis>=5.4.0
django-webpack-loader>=3.0.0
djangorestframework>=3.14.0
docutils>=0.20.1
h11>=0.14.0
h2>=4.1.0
hpack>=4.0.0
httptools>=0.6.1
hyperframe>=6.0.1
hyperlink>=21.0.0
idna>=3.6
incremental>=22.10.0
markdown-it-py>=3.0.0
mdurl>=0.1.2
msgpack>=1.0.7
oauthlib>=3.2.2
Pillow>=10.1.0
priority>=2.0.0
psycopg>=3.1.16
psycopg-binary>=3.1.16
psycopg[c]>=3.1.16
psycopg-pool>=3.2.0
pyasn1>=0.5.1
pyasn1-modules>=0.3.0
pycparser>=2.21
Pygments>=2.17.2
PyJWT>=2.8.0
pyOpenSSL>=23.3.0
python-dotenv>=1.0.0
pytz>=2023.3.post1
PyYAML>=5.3.1
redis>=5.1.0a1
requests>=2.31.0
requests-oauthlib>=1.3.1
rich>=13.7.0
service-identity>=23.1.0
six>=1.16.0
sniffio>=1.3.0
soupsieve>=2.5
sqlparse>=0.4.4
Twisted>=23.10.0
txaio>=23.1.1
typing_extensions>=4.9.0
urllib3>=2.1.0
uvicorn>=0.25.0
uvloop>=0.19.0
watchfiles>=0.21.0
websockets>=12.0
whitenoise>=6.6.0
zope.interface>=6.1
graphene-django>=3.0.0
channels_redis~=4.2.1
channels>=4.1.0
adrf~=0.1.8
trio~=0.27.0
docker~=7.1.0
Hypercorn~=0.17.3
drf-spectacular~=0.28.0
diffusers~=0.32.1
huggingface-hub~=0.27.0
transformers~=4.47.1

Jinja2~=3.1.4